refactor(llm): extract request/response logging into _LoggingMixin
client.py was 890 lines; ~200 of them were the full-fidelity request/
response logging (_log_request with incremental message + system-prompt
diffing, _log_response). That formatting logic is orthogonal to the
chat / retry / streaming machinery, so move it into a _LoggingMixin
(agentao/llm/_logging.py) following the established mixin pattern.

LLMClient(_LoggingMixin) keeps both methods as instance methods, so the
test that calls c._log_request(...) directly and every internal
self._log_request/_log_response call work unchanged. Method bodies moved
verbatim (semantic diff confirms no logic change). Dropped the now-unused
`import json` from client.py. client.py: 890 → 688 lines. Full suite
green (2840 passed).
